LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/reporting/cross-origin-reports-isolated.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://127.0.0.1:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL https://www.localhost:9443/reporting/resources/report.py?reportID=GENERATED_REPORT_ID
CONSOLE MESSAGE: Refused to load https://localhost:9443/reporting/resources/fail.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL https://www.localhost:9443/reporting/resources/report.py?reportID=GENERATED_REPORT_ID


FAIL Reports were sent in two requests. assert_equals: expected 2 but got 0
